driver-drowsiness-web/server/requirements.txt
opencv-python>=4.5.0
dlib>=19.22.0
imutils>=0.5.4
scipy>=1.7.0
numpy>=1.20.0
python-socketio[client]>=5.7.2 